feat(cli): Add command-line tool that wraps generation + bindata
This commit adds a command-line tool that wraps generating the swagger.json
along with embedding it in a Go file. This allows consumers to install the
package and add the following to their source:

//go:generate mkswagger -i github.com/lob/my_service/cmd/entry_pkg

The resulting swagger.json and bindata.go will be placed in the same
package and directory where the generate command appears.

This has a few advantages over the previous approach:

* It requires the consumer to install fewer tools
* It works correctly with vendored packages; previously `go generate` had
  to be invoked with the aboslute disk path to the magic `go-swagger-tools/generate`
  package.
* It utilizes the `go generate` framework in a way that's consisten with
  other tools: the user invokes `go generate` on _their_ code, not the
  tooling package.
* It doesn't rely on magic environment variables.