fix(installation): Be more precise with calls into go-swagger. (#3)
Previously when attempting to install via `go get`, the go compiler
would report a type error. This was due to the fact that `go-swagger`
includes a `vendor` directory, and expects to be used as a utility, not
as a library. Go prevents explicit imports from `vendor` paths, so
instead we make a more precise call into the `go-swagger` machinery,
calling the scanner and marshalling the result to JSON ourselves to
avoid writing the type.

/shrug #yolo